Rename pkg verifyhandler to verify
