+++ cmd/auth/main.go
@@ -18,7 +18,7 @@ import (
 	"github.com/skygeario/skygear-server/pkg/auth/handler"
 	forgotpwdhandler "github.com/skygeario/skygear-server/pkg/auth/handler/forgotpwd"
 	"github.com/skygeario/skygear-server/pkg/auth/handler/ssohandler"
-	"github.com/skygeario/skygear-server/pkg/auth/handler/verifyhandler"
+	userverifyhandler "github.com/skygeario/skygear-server/pkg/auth/handler/userverify"
 	"github.com/skygeario/skygear-server/pkg/auth/task"
 	"github.com/skygeario/skygear-server/pkg/core/config"
 	"github.com/skygeario/skygear-server/pkg/core/logging"
@@ -77,8 +77,8 @@ func main() {
 	handler.AttachWelcomeEmailHandler(&srv, authDependency)
 	forgotpwdhandler.AttachForgotPasswordHandler(&srv, authDependency)
 	forgotpwdhandler.AttachForgotPasswordResetHandler(&srv, authDependency)
-	verifyhandler.AttachVerifyRequestHandler(&srv, authDependency)
-	verifyhandler.AttachVerifyCodeHandler(&srv, authDependency)
+	userverifyhandler.AttachVerifyRequestHandler(&srv, authDependency)
+	userverifyhandler.AttachVerifyCodeHandler(&srv, authDependency)
 	ssohandler.AttachAuthURLHandler(&srv, authDependency)
 	ssohandler.AttachConfigHandler(&srv, authDependency)
 	ssohandler.AttachIFrameHandlerFactory(&srv, authDependency)
