+++ pkg/auth/handler/userverify/handler.go
@@ -1,4 +1,4 @@
-package verifyhandler
+package userverify
 
 import (
 	"time"
