+++ pkg/auth/handler/userverify/verify_code.go
@@ -1,4 +1,4 @@
-package verifyhandler
+package userverify
 
 import (
 	"encoding/json"
