+++ pkg/auth/handler/userverify/verify_code_test.go
@@ -1,4 +1,4 @@
-package verifyhandler
+package userverify
 
 import (
 	"net/http"
